Retain shipped_at date when a load is unshipped
When a load is shipped:

If shipped_at is already set don't update it on the load or pallets.
If shipped_at is null, set it to Time.now on the load and pallets (current behaviour)
When a load is unshipped:

Unset the shipped boolean on the load, but leave shipped_at as-is on the load and pallets
When a pallet is de-allocated from a load:

Set shipped_at to null on the pallet.