Hi Lux
Like the new/edit, but showing labels and making non-applicable columns invisible.
And at the bottom, a list of the targets in the following format:

target: directory/email addresses
Where target is paltrack/mail:xyz and directory is the full path as calculated from the config file, and email addresses is the list of mail recipients for the applicable target.

Use EdiOutRepo.load_config and look in BaseEdiOutConfig for ways to access the dir/mail addresses.

Also you can display an active boolean which shows true if the edi_out_config entry for the flow type is true under :send_edi. (note the flow type there is a lowercase symbol).